pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "url-to-pdf"
version = "0.1.0"
description = "CLI tool to extract article content from URL and save as PDF"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Roman Averin" }]
dependencies = [
  "newspaper4k>=0.9.0",
  "fpdf2>=2.7.0",
  "click>=8.0.0",
  "requests>=2.28.0",
  "Pillow>=9.0.0",
  "lxml>=4.9.0",
  "lxml_html_clean>=0.1.0",
  "rich>=13.0.0",
]

[project.scripts]
url-to-pdf = "url_to_pdf.cli:main"

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-cov>=4.0.0", "pylint>=3.0.0"]

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v"

[tool.pylint.main]
py-version = "3.10"
jobs = 0
load-plugins = []

[tool.pylint.messages_control]
disable = [
  "missing-module-docstring",
  "missing-class-docstring",
  "missing-function-docstring",
  "too-few-public-methods",
  "line-too-long",
  "too-many-positional-arguments",
  "too-many-instance-attributes",
  "too-many-locals",
  "too-many-branches",
  "broad-exception-caught",
]

[tool.pylint.format]
max-line-length = 100

[tool.pylint.design]
max-args = 10

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "ruff>=0.14.13",
]
